fix compilation error and move utility methods in BatchPhysicalRel to FlinkRelMdUtil
